+++ argon-dashboard-master/examples/meja.php
@@ -1,6 +1,20 @@
 <?php 
-  require 'functions.php';
-  $meja = query("SELECT * FROM meja");
+session_start();
+
+if (!isset($_SESSION["username"])) {
+  header("Location: login.php");
+  exit();
+}
+
+require 'functions.php';
+$meja = query("SELECT * FROM meja");
+
+// ketika tombol cari diklik
+if (isset($_POST['carimeja'])) {
+  $meja = carimeja($_POST['keyword']);
+}
+
+
 ?>
 <!--
 =========================================================
@@ -98,10 +112,24 @@
               </a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="pemasukan_pengeluaran.php">
+              <a class="nav-link" data-toggle="collapse" href="#dua">
                 <i class="ni ni-bullet-list-67 text-primary"></i>
                 <span class="nav-link-text">Pemasukan dan Pengeluaran</span>
               </a>
+              <div class="collapse" id="dua">
+                <ul class="nav nav-collapse">
+                  <li class="nav-item">
+                    <a href="pemasukan.php" class="nav-link">
+                      <span class="nav-link-text">Data Pemasukan</span>
+                    </a>
+                  </li>
+                  <li class="nav-item">
+                    <a href="pengeluaran.php" class="nav-link">
+                      <span class="nav-link-text">Data Pengeluaran</span>
+                    </a>
+                  </li>
+                </ul>
+              </div>
             </li>
             <li class="nav-item">
               <a class="nav-link" data-toggle="collapse" href="#tables">
@@ -146,8 +174,8 @@
                     </a>
                   </li>
                   <li class="nav-item">
-                    <a href="laporanomset.php" class="nav-link">
-                      <span class="nav-link-text">Data Omset</span>
+                    <a href="laporankeuntungan.php" class="nav-link">
+                      <span class="nav-link-text">Data Keuntungan</span>
                     </a>
                   </li>
                 </ul>
@@ -167,13 +195,14 @@
       <div class="container-fluid">
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
           <!-- Search form -->
-          <form class="navbar-search navbar-search-light form-inline mr-sm-3" id="navbar-search-main">
+          <form class="navbar-search navbar-search-light form-inline mr-sm-3" id="navbar-search-main" action="" method="POST">
             <div class="form-group mb-0">
               <div class="input-group input-group-alternative input-group-merge">
                 <div class="input-group-prepend">
                   <span class="input-group-text"><i class="fas fa-search"></i></span>
                 </div>
-                <input class="form-control" placeholder="Search" type="text">
+                <input class="keyword form-control" placeholder="Search" type="text" name="keyword" autocomplete="off">
+                <button type="submit" name="carimeja" class="tombol-cari btn"></button>
               </div>
             </div>
             <button type="button" class="close" data-action="search-close" data-target="#navbar-search-main" aria-label="Close">
@@ -246,7 +275,7 @@
                     <img alt="Image placeholder" src="../assets/img/theme/team-4.jpg">
                   </span>
                   <div class="media-body  ml-2  d-none d-lg-block">
-                    <span class="mb-0 text-sm  font-weight-bold">John Snow</span>
+                    <span class="mb-0 text-sm  font-weight-bold"><?php echo $_SESSION["username"]; ?></span>
                   </div>
                 </div>
               </a>
@@ -263,7 +292,7 @@
                   <span>Settings</span>
                 </a>
                 <div class="dropdown-divider"></div>
-                <a href="#!" class="dropdown-item">
+                <a href="logout.php" onclick="return confirm('Anda yakin akan keluar?')" class="dropdown-item">
                   <i class="ni ni-user-run"></i>
                   <span>Logout</span>
                 </a>
@@ -291,7 +320,7 @@
                   <div class="card-header bg-transparent border-0">
                     <h3 class="text-white mb-0">Meja</h3>
                   </div>
-                  <div class="table-responsive">
+                  <div class="hasil-cari table-responsive">
                     <table class="table align-items-center table-dark table-flush">
                       <thead class="thead-dark">
                         <tr>
@@ -309,8 +338,8 @@
                             <td><?php echo $mj['id_meja'];?></td>
                             <td><?php echo $mj['status'];?></td>
                             <td>
-                            <a href="ubahmeja.php?id_meja=<?php echo $mj['id_meja']; ?>"><i class="far fa-edit text-white"></i></a> |
-                            <a href="hapusmeja.php?id_meja=<?php echo $mj['id_meja']; ?>"><i class="far fa-trash-alt text-white"></i></a>
+                            <a href="ubahmeja.php?id_meja=<?php echo $mj['id_meja']; ?>" onclick="return confirm('Ubah data meja?')"><i class="far fa-edit text-white"></i></a> |
+                            <a href="hapusmeja.php?id_meja=<?php echo $mj['id_meja']; ?>" onclick="return confirm('Hapus data meja?')"><i class="far fa-trash-alt text-white"></i></a>
                             </td>
                           </tr>
                           <?php } ?>
@@ -349,6 +378,8 @@
   <script src="../assets/vendor/clipboard/dist/clipboard.min.js"></script>
   <!-- Argon JS -->
   <script src="../assets/js/argon.js?v=1.2.0"></script>
+  <!-- my js -->
+  <script src="../assets/js/script_meja.js"></script>
 </body>
 
 </html>
